backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
redis==5.0.1
bcrypt==4.0.1
alembic==1.13.0
httpx==0.25.2
jinja2==3.1.2
python-dotenv==1.0.0